move gw plotting to plot_tools
